refactor(SE-60): enhance readability of ResetPasswordServiceImpl
Refactored ResetPasswordServiceImpl to improve code clarity and maintainability.
